refactor: group EditGameServer inputs and form sections into subfolders
