+++ src/components/display/GameServer/EditGameServer/inputs/ListInputEditGameServer.tsx
